Break resource cleaner into separate packages
