+++ test/hack/cleanup/main.go
@@ -22,48 +22,16 @@ import (
 
 	"github.com/aws/aws-sdk-go-v2/config"
 	"github.com/aws/aws-sdk-go-v2/service/cloudformation"
-	cloudformationtypes "github.com/aws/aws-sdk-go-v2/service/cloudformation/types"
 	"github.com/aws/aws-sdk-go-v2/service/ec2"
-	ec2types "github.com/aws/aws-sdk-go-v2/service/ec2/types"
-
 	"github.com/aws/aws-sdk-go-v2/service/iam"
-	"github.com/aws/aws-sdk-go-v2/service/timestreamwrite"
-	timestreamtypes "github.com/aws/aws-sdk-go-v2/service/timestreamwrite/types"
-	"github.com/aws/aws-sdk-go/aws"
 	"github.com/samber/lo"
-	"go.uber.org/multierr"
 	"go.uber.org/zap"
-	"k8s.io/client-go/util/workqueue"
+
+	"github.com/aws/karpenter/test/hack/cleanup/metrics"
+	"github.com/aws/karpenter/test/hack/cleanup/resource"
 )
 
-const (
-	expirationTTL            = time.Hour * 12
-	karpenterMetricRegion    = "us-east-2"
-	karpenterMetricDatabase  = "karpenterTesting"
-	karpenterMetricTableName = "sweeperCleanedResources"
-
-	karpenterClusterNameTag     = "karpenter.sh/managed-by"
-	karpenterProvisionerNameTag = "karpenter.sh/provisioner-name"
-	karpenterNodePoolTag        = "karpenter.sh/nodepool"
-	karpenterLaunchTemplateTag  = "karpenter.k8s.aws/cluster"
-	karpenterSecurityGroupTag   = "karpenter.sh/discovery"
-	// TODO @joinnis: Remove this karpenterTestingTagLegacy field after running this cleanup script for a few days
-	karpenterTestingTagLegacy = "testing.karpenter.sh/cluster"
-	karpenterTestingTag       = "testing/cluster"
-	k8sClusterTag             = "cluster.k8s.amazonaws.com/name"
-	githubRunURLTag           = "github.com/run-url"
-)
-
-type CleanableResourceType interface {
-	Type() string
-	GetExpired(context.Context, time.Time) ([]string, error)
-	Get(context.Context, string) ([]string, error)
-	Cleanup(context.Context, []string) ([]string, error)
-}
-
-type MetricsClient interface {
-	FireMetric(context.Context, string, float64, string) error
-}
+const expirationTTL = time.Hour * 12
 
 func main() {
 	var clusterName string
@@ -83,19 +51,24 @@ func main() {
 	cloudFormationClient := cloudformation.NewFromConfig(cfg)
 	iamClient := iam.NewFromConfig(cfg)
 
-	metricsClient := MetricsClient(&timestream{timestreamClient: timestreamwrite.NewFromConfig(cfg, WithRegion(karpenterMetricRegion))})
+	metricsClient := metrics.Client(metrics.NewTimeStream(cfg))
 
-	resources := []CleanableResourceType{
-		&eni{ec2Client: ec2Client},
-		&securitygroup{ec2Client: ec2Client},
-		&instance{ec2Client: ec2Client},
-		&stack{cloudFormationClient: cloudFormationClient},
-		&launchtemplate{ec2Client: ec2Client},
-		&oidc{iamClient: iamClient},
-		&instanceProfile{iamClient: iamClient},
+	// These resources are intentionally ordered so that instances that are using ENIs
+	// will be cleaned before ENIs are attempted to be cleaned up. Likewise, instances and ENIs
+	// are cleaned up before security groups are cleaned up to ensure that everything is detached and doesn't
+	// prevent deletion
+	resources := []resource.Resource{
+		resource.NewInstance(ec2Client),
+		resource.NewENI(ec2Client),
+		resource.NewSecurityGroup(ec2Client),
+		resource.NewLaunchTemplate(ec2Client),
+		resource.NewOIDC(iamClient),
+		resource.NewInstanceProfile(iamClient),
+		resource.NewStack(cloudFormationClient),
 	}
 
-	workqueue.ParallelizeUntil(ctx, len(resources), len(resources), func(i int) {
+	for i := range resources {
+		resourceLogger := logger.With("type", resources[i].Type())
 		var ids []string
 		var err error
 		if clusterName == "" {
@@ -104,610 +77,18 @@ func main() {
 			ids, err = resources[i].Get(ctx, clusterName)
 		}
 		if err != nil {
-			logger.With("type", resources[i].Type()).Errorf("%v", err)
+			resourceLogger.Errorf("%v", err)
 		}
-		logger.With("type", resources[i].Type(), "ids", ids, "count", len(ids)).Infof("discovered resources")
+		resourceLogger.With("ids", ids, "count", len(ids)).Infof("discovered resources")
 		if len(ids) > 0 {
 			cleaned, err := resources[i].Cleanup(ctx, ids)
 			if err != nil {
-				logger.With("type", resources[i].Type()).Errorf("%v", err)
+				resourceLogger.Errorf("%v", err)
 			}
 			if err = metricsClient.FireMetric(ctx, fmt.Sprintf("%sDeleted", resources[i].Type()), float64(len(cleaned)), cfg.Region); err != nil {
-				logger.With("type", resources[i].Type()).Errorf("%v", err)
+				resourceLogger.Errorf("%v", err)
 			}
-			logger.With("type", resources[i].Type(), "ids", cleaned, "count", len(cleaned)).Infof("deleted resources")
+			resourceLogger.With("ids", cleaned, "count", len(cleaned)).Infof("deleted resources")
 		}
-	})
-}
-
-type instance struct {
-	ec2Client *ec2.Client
-}
-
-func (i *instance) Type() string {
-	return "Instances"
-}
-
-func (i *instance) GetExpired(ctx context.Context, expirationTime time.Time) (ids []string, err error) {
-	var nextToken *string
-	for {
-		out, err := i.ec2Client.DescribeInstances(ctx, &ec2.DescribeInstancesInput{
-			Filters: []ec2types.Filter{
-				{
-					Name:   lo.ToPtr("instance-state-name"),
-					Values: []string{string(ec2types.InstanceStateNameRunning)},
-				},
-				{
-					Name:   lo.ToPtr("tag-key"),
-					Values: []string{karpenterProvisionerNameTag, karpenterNodePoolTag},
-				},
-			},
-			NextToken: nextToken,
-		})
-		if err != nil {
-			return ids, err
-		}
-
-		for _, res := range out.Reservations {
-			for _, instance := range res.Instances {
-				if lo.FromPtr(instance.LaunchTime).Before(expirationTime) {
-					ids = append(ids, lo.FromPtr(instance.InstanceId))
-				}
-			}
-		}
-
-		nextToken = out.NextToken
-		if nextToken == nil {
-			break
-		}
-	}
-	return ids, err
-}
-
-func (i *instance) Get(ctx context.Context, clusterName string) (ids []string, err error) {
-	var nextToken *string
-
-	for {
-		out, err := i.ec2Client.DescribeInstances(ctx, &ec2.DescribeInstancesInput{
-			Filters: []ec2types.Filter{
-				{
-					Name:   lo.ToPtr("instance-state-name"),
-					Values: []string{string(ec2types.InstanceStateNameRunning)},
-				},
-				{
-					Name:   lo.ToPtr("tag:" + karpenterClusterNameTag),
-					Values: []string{clusterName},
-				},
-			},
-			NextToken: nextToken,
-		})
-		if err != nil {
-			return ids, err
-		}
-
-		for _, res := range out.Reservations {
-			for _, instance := range res.Instances {
-				ids = append(ids, lo.FromPtr(instance.InstanceId))
-			}
-		}
-
-		nextToken = out.NextToken
-		if nextToken == nil {
-			break
-		}
-	}
-	return ids, err
-}
-
-// Terminate any old instances that were provisioned by Karpenter as part of testing
-// We execute these in serial since we will most likely get rate limited if we try to delete these too aggressively
-func (i *instance) Cleanup(ctx context.Context, ids []string) ([]string, error) {
-	if _, err := i.ec2Client.TerminateInstances(ctx, &ec2.TerminateInstancesInput{
-		InstanceIds: ids,
-	}); err != nil {
-		return nil, err
-	}
-	return ids, nil
-}
-
-type securitygroup struct {
-	ec2Client *ec2.Client
-}
-
-func (sg *securitygroup) Type() string {
-	return "SecurityGroup"
-}
-
-func (sg *securitygroup) GetExpired(ctx context.Context, expirationTime time.Time) (ids []string, err error) {
-	var nextToken *string
-	for {
-		out, err := sg.ec2Client.DescribeSecurityGroups(ctx, &ec2.DescribeSecurityGroupsInput{
-			Filters: []ec2types.Filter{
-				{
-					Name:   lo.ToPtr("group-name"),
-					Values: []string{"security-group-drift"},
-				},
-			},
-			NextToken: nextToken,
-		})
-		if err != nil {
-			return ids, err
-		}
-
-		for _, sgroup := range out.SecurityGroups {
-			creationDate, found := lo.Find(sgroup.Tags, func(tag ec2types.Tag) bool {
-				return *tag.Key == "creation-date"
-			})
-			if !found {
-				continue
-			}
-			time, err := time.Parse(time.RFC3339, *creationDate.Value)
-			if err != nil {
-				continue
-			}
-			if time.Before(expirationTime) {
-				ids = append(ids, lo.FromPtr(sgroup.GroupId))
-			}
-		}
-
-		nextToken = out.NextToken
-		if nextToken == nil {
-			break
-		}
-	}
-	return ids, err
-}
-
-func (sg *securitygroup) Get(ctx context.Context, clusterName string) (ids []string, err error) {
-	var nextToken *string
-
-	for {
-		out, err := sg.ec2Client.DescribeSecurityGroups(ctx, &ec2.DescribeSecurityGroupsInput{
-			Filters: []ec2types.Filter{
-				{
-					Name:   lo.ToPtr("tag:" + karpenterSecurityGroupTag),
-					Values: []string{clusterName},
-				},
-			},
-			NextToken: nextToken,
-		})
-		if err != nil {
-			return ids, err
-		}
-
-		for _, sgroup := range out.SecurityGroups {
-			ids = append(ids, lo.FromPtr(sgroup.GroupId))
-		}
-
-		nextToken = out.NextToken
-		if nextToken == nil {
-			break
-		}
-	}
-	return ids, err
-}
-
-func (sg *securitygroup) Cleanup(ctx context.Context, ids []string) ([]string, error) {
-	deleted := []string{}
-	var errs error
-	for i := range ids {
-		_, err := sg.ec2Client.DeleteSecurityGroup(ctx, &ec2.DeleteSecurityGroupInput{
-			GroupId: aws.String(ids[i]),
-		})
-		if err != nil {
-			errs = multierr.Append(errs, err)
-			continue
-		}
-		deleted = append(deleted, ids[i])
-	}
-
-	return deleted, errs
-}
-
-type stack struct {
-	cloudFormationClient *cloudformation.Client
-}
-
-func (s *stack) Type() string {
-	return "CloudformationStacks"
-}
-
-func (s *stack) GetExpired(ctx context.Context, expirationTime time.Time) (names []string, err error) {
-	var nextToken *string
-	for {
-		out, err := s.cloudFormationClient.DescribeStacks(ctx, &cloudformation.DescribeStacksInput{
-			NextToken: nextToken,
-		})
-		if err != nil {
-			return names, err
-		}
-
-		stacks := lo.Reject(out.Stacks, func(s cloudformationtypes.Stack, _ int) bool {
-			return s.StackStatus == cloudformationtypes.StackStatusDeleteComplete ||
-				s.StackStatus == cloudformationtypes.StackStatusDeleteInProgress
-		})
-		for _, stack := range stacks {
-			if _, found := lo.Find(stack.Tags, func(t cloudformationtypes.Tag) bool {
-				return lo.FromPtr(t.Key) == githubRunURLTag
-			}); found && lo.FromPtr(stack.CreationTime).Before(expirationTime) {
-				names = append(names, lo.FromPtr(stack.StackName))
-			}
-		}
-
-		nextToken = out.NextToken
-		if nextToken == nil {
-			break
-		}
-	}
-	return names, err
-}
-
-func (s *stack) Get(ctx context.Context, clusterName string) (names []string, err error) {
-	return []string{fmt.Sprintf("iam-%s", clusterName), fmt.Sprintf("eksctl-%s-cluster", clusterName)}, nil
-}
-
-// Terminate any old stacks that were provisioned as part of testing
-// We execute these in serial since we will most likely get rate limited if we try to delete these too aggressively
-func (s *stack) Cleanup(ctx context.Context, names []string) ([]string, error) {
-	var errs error
-	deleted := []string{}
-	for i := range names {
-		_, err := s.cloudFormationClient.DeleteStack(ctx, &cloudformation.DeleteStackInput{
-			StackName: lo.ToPtr(names[i]),
-		})
-		if err != nil {
-			errs = multierr.Append(errs, err)
-			continue
-		}
-		deleted = append(deleted, names[i])
-	}
-	return deleted, errs
-}
-
-type launchtemplate struct {
-	ec2Client *ec2.Client
-}
-
-func (lt *launchtemplate) Type() string {
-	return "LaunchTemplates"
-}
-
-func (lt *launchtemplate) GetExpired(ctx context.Context, expirationTime time.Time) (names []string, err error) {
-	var nextToken *string
-	for {
-		out, err := lt.ec2Client.DescribeLaunchTemplates(ctx, &ec2.DescribeLaunchTemplatesInput{
-			Filters: []ec2types.Filter{
-				{
-					Name:   lo.ToPtr("tag-key"),
-					Values: []string{karpenterLaunchTemplateTag},
-				},
-			},
-			NextToken: nextToken,
-		})
-		if err != nil {
-			return names, err
-		}
-
-		for _, launchTemplate := range out.LaunchTemplates {
-			if lo.FromPtr(launchTemplate.CreateTime).Before(expirationTime) {
-				names = append(names, lo.FromPtr(launchTemplate.LaunchTemplateName))
-			}
-		}
-
-		nextToken = out.NextToken
-		if nextToken == nil {
-			break
-		}
-	}
-	return names, err
-}
-
-func (lt *launchtemplate) Get(ctx context.Context, clusterName string) (names []string, err error) {
-	var nextToken *string
-	for {
-		out, err := lt.ec2Client.DescribeLaunchTemplates(ctx, &ec2.DescribeLaunchTemplatesInput{
-			Filters: []ec2types.Filter{
-				{
-					Name:   lo.ToPtr("tag:" + karpenterLaunchTemplateTag),
-					Values: []string{clusterName},
-				},
-			},
-			NextToken: nextToken,
-		})
-		if err != nil {
-			return names, err
-		}
-
-		for _, launchTemplate := range out.LaunchTemplates {
-			names = append(names, lo.FromPtr(launchTemplate.LaunchTemplateName))
-		}
-
-		nextToken = out.NextToken
-		if nextToken == nil {
-			break
-		}
-	}
-	return names, err
-}
-
-// Terminate any old launch templates that were managed by Karpenter and were provisioned as part of testing
-// We execute these in serial since we will most likely get rate limited if we try to delete these too aggressively
-func (lt *launchtemplate) Cleanup(ctx context.Context, names []string) ([]string, error) {
-	var errs error
-	deleted := []string{}
-	for i := range names {
-		_, err := lt.ec2Client.DeleteLaunchTemplate(ctx, &ec2.DeleteLaunchTemplateInput{
-			LaunchTemplateName: lo.ToPtr(names[i]),
-		})
-		if err != nil {
-			errs = multierr.Append(errs, err)
-			continue
-		}
-		deleted = append(deleted, names[i])
-	}
-	return deleted, errs
-}
-
-type oidc struct {
-	iamClient *iam.Client
-}
-
-func (o *oidc) Type() string {
-	return "OpenIDConnectProvider"
-}
-
-func (o *oidc) GetExpired(ctx context.Context, expirationTime time.Time) (names []string, err error) {
-	out, err := o.iamClient.ListOpenIDConnectProviders(ctx, &iam.ListOpenIDConnectProvidersInput{})
-	if err != nil {
-		return names, err
-	}
-
-	errs := make([]error, len(out.OpenIDConnectProviderList))
-	for i := range out.OpenIDConnectProviderList {
-		oicd, err := o.iamClient.GetOpenIDConnectProvider(ctx, &iam.GetOpenIDConnectProviderInput{
-			OpenIDConnectProviderArn: out.OpenIDConnectProviderList[i].Arn,
-		})
-		if err != nil {
-			errs[i] = err
-			continue
-		}
-
-		for _, t := range oicd.Tags {
-			if lo.FromPtr(t.Key) == githubRunURLTag && oicd.CreateDate.Before(expirationTime) {
-				names = append(names, lo.FromPtr(out.OpenIDConnectProviderList[i].Arn))
-			}
-		}
-	}
-
-	return names, multierr.Combine(errs...)
-}
-
-func (o *oidc) Get(ctx context.Context, clusterName string) (names []string, err error) {
-	return names, err
-}
-
-// Cleanup any old OIDC providers that were are remaining as part of testing
-// We execute these in serial since we will most likely get rate limited if we try to delete these too aggressively
-func (o *oidc) Cleanup(ctx context.Context, arns []string) ([]string, error) {
-	var errs error
-	deleted := []string{}
-	for i := range arns {
-		_, err := o.iamClient.DeleteOpenIDConnectProvider(ctx, &iam.DeleteOpenIDConnectProviderInput{
-			OpenIDConnectProviderArn: lo.ToPtr(arns[i]),
-		})
-		if err != nil {
-			errs = multierr.Append(errs, err)
-			continue
-		}
-		deleted = append(deleted, arns[i])
-	}
-	return deleted, errs
-}
-
-type instanceProfile struct {
-	iamClient *iam.Client
-}
-
-func (ip *instanceProfile) Type() string {
-	return "InstanceProfile"
-}
-
-func (ip *instanceProfile) GetExpired(ctx context.Context, expirationTime time.Time) (names []string, err error) {
-	out, err := ip.iamClient.ListInstanceProfiles(ctx, &iam.ListInstanceProfilesInput{})
-	if err != nil {
-		return names, err
-	}
-
-	errs := make([]error, len(out.InstanceProfiles))
-	for i := range out.InstanceProfiles {
-		profiles, err := ip.iamClient.ListInstanceProfileTags(ctx, &iam.ListInstanceProfileTagsInput{
-			InstanceProfileName: out.InstanceProfiles[i].InstanceProfileName,
-		})
-		if err != nil {
-			errs[i] = err
-			continue
-		}
-
-		for _, t := range profiles.Tags {
-			// Since we can only get the date of the instance profile (not the exact time the instance profile was created)
-			// we add a day to the time that it was created to account for the worst-case of the instance profile being created
-			// at 23:59:59 and being marked with a time of 00:00:00 due to only capturing the date and not the time
-			if lo.FromPtr(t.Key) == karpenterTestingTag && out.InstanceProfiles[i].CreateDate.Add(time.Hour*24).Before(expirationTime) {
-				names = append(names, lo.FromPtr(out.InstanceProfiles[i].InstanceProfileName))
-			}
-		}
-	}
-
-	return names, multierr.Combine(errs...)
-}
-
-func (ip *instanceProfile) Get(ctx context.Context, clusterName string) (names []string, err error) {
-	out, err := ip.iamClient.ListInstanceProfiles(ctx, &iam.ListInstanceProfilesInput{})
-	if err != nil {
-		return names, err
-	}
-
-	errs := make([]error, len(out.InstanceProfiles))
-	for i := range out.InstanceProfiles {
-		profiles, err := ip.iamClient.ListInstanceProfileTags(ctx, &iam.ListInstanceProfileTagsInput{
-			InstanceProfileName: out.InstanceProfiles[i].InstanceProfileName,
-		})
-		if err != nil {
-			errs[i] = err
-			continue
-		}
-
-		for _, t := range profiles.Tags {
-			if lo.FromPtr(t.Key) == karpenterTestingTag && lo.FromPtr(t.Value) == clusterName {
-				names = append(names, lo.FromPtr(out.InstanceProfiles[i].InstanceProfileName))
-			}
-		}
-	}
-
-	return names, multierr.Combine(errs...)
-}
-
-func (ip *instanceProfile) Cleanup(ctx context.Context, names []string) ([]string, error) {
-	var errs error
-	deleted := []string{}
-	for i := range names {
-		out, _ := ip.iamClient.GetInstanceProfile(ctx, &iam.GetInstanceProfileInput{InstanceProfileName: lo.ToPtr(names[i])})
-		if len(out.InstanceProfile.Roles) != 0 {
-			_, _ = ip.iamClient.RemoveRoleFromInstanceProfile(ctx, &iam.RemoveRoleFromInstanceProfileInput{
-				InstanceProfileName: lo.ToPtr(names[i]),
-				RoleName:            out.InstanceProfile.Roles[0].RoleName,
-			})
-		}
-		_, err := ip.iamClient.DeleteInstanceProfile(ctx, &iam.DeleteInstanceProfileInput{
-			InstanceProfileName: lo.ToPtr(names[i]),
-		})
-		if err != nil {
-			errs = multierr.Append(errs, err)
-		}
-	}
-	return deleted, errs
-}
-
-type eni struct {
-	ec2Client *ec2.Client
-}
-
-func (e *eni) Type() string {
-	return "ElasticNetworkInterface"
-}
-
-func (e *eni) GetExpired(ctx context.Context, expirationTime time.Time) (ids []string, err error) {
-	var nextToken *string
-	for {
-		out, err := e.ec2Client.DescribeNetworkInterfaces(ctx, &ec2.DescribeNetworkInterfacesInput{
-			Filters: []ec2types.Filter{
-				{
-					Name:   lo.ToPtr("tag-key"),
-					Values: []string{k8sClusterTag},
-				},
-			},
-			NextToken: nextToken,
-		})
-		if err != nil {
-			return ids, err
-		}
-
-		for _, ni := range out.NetworkInterfaces {
-			creationDate, found := lo.Find(ni.TagSet, func(tag ec2types.Tag) bool {
-				return *tag.Key == "node.k8s.amazonaws.com/createdAt"
-			})
-			if !found {
-				continue
-			}
-			time, err := time.Parse(time.RFC3339, *creationDate.Value)
-			if err != nil {
-				continue
-			}
-			if ni.Status == ec2types.NetworkInterfaceStatusAvailable && time.Before(expirationTime) {
-				ids = append(ids, lo.FromPtr(ni.NetworkInterfaceId))
-			}
-		}
-
-		nextToken = out.NextToken
-		if nextToken == nil {
-			break
-		}
-	}
-	return ids, err
-}
-
-func (e *eni) Get(ctx context.Context, clusterName string) (ids []string, err error) {
-	var nextToken *string
-	for {
-		out, err := e.ec2Client.DescribeNetworkInterfaces(ctx, &ec2.DescribeNetworkInterfacesInput{
-			Filters: []ec2types.Filter{
-				{
-					Name:   lo.ToPtr("tag:" + k8sClusterTag),
-					Values: []string{clusterName},
-				},
-			},
-			NextToken: nextToken,
-		})
-		if err != nil {
-			return ids, err
-		}
-
-		for _, ni := range out.NetworkInterfaces {
-			ids = append(ids, lo.FromPtr(ni.NetworkInterfaceId))
-		}
-
-		nextToken = out.NextToken
-		if nextToken == nil {
-			break
-		}
-	}
-	return ids, err
-}
-
-func (e *eni) Cleanup(ctx context.Context, ids []string) ([]string, error) {
-	deleted := []string{}
-	var errs error
-	for i := range ids {
-		_, err := e.ec2Client.DeleteNetworkInterface(ctx, &ec2.DeleteNetworkInterfaceInput{
-			NetworkInterfaceId: aws.String(ids[i]),
-		})
-		if err != nil {
-			errs = multierr.Append(errs, err)
-			continue
-		}
-		deleted = append(deleted, ids[i])
-	}
-
-	return deleted, errs
-}
-
-type timestream struct {
-	timestreamClient *timestreamwrite.Client
-}
-
-func (t *timestream) FireMetric(ctx context.Context, name string, value float64, region string) error {
-	_, err := t.timestreamClient.WriteRecords(ctx, &timestreamwrite.WriteRecordsInput{
-		DatabaseName: aws.String(karpenterMetricDatabase),
-		TableName:    aws.String(karpenterMetricTableName),
-		Records: []timestreamtypes.Record{
-			{
-				MeasureName:  aws.String(name),
-				MeasureValue: aws.String(fmt.Sprintf("%f", value)),
-				Time:         aws.String(fmt.Sprintf("%d", time.Now().UnixMilli())),
-				Dimensions: []timestreamtypes.Dimension{
-					{
-						Name:  aws.String("region"),
-						Value: aws.String(region),
-					},
-				},
-			},
-		},
-	})
-	return err
-}
-
-func WithRegion(region string) func(*timestreamwrite.Options) {
-	return func(o *timestreamwrite.Options) {
-		o.Region = region
 	}
 }
